Working on product admin page

+++ Backend/controllers/subCategory.controller.js
@@ -1,3 +1,4 @@
+import ProductModel from "../models/product.model.js";
 import SubCategoryModel from "../models/subCategory.model.js";
 
 export const AddSubCategoryController = async (req, res) => {
@@ -94,6 +95,20 @@ export const updateSubCategoryController = async (req, res) => {
 export const deleteSubCategoryController = async (req, res) => {
   try {
     const {_id} = req.body
+
+    const checkProduct = await ProductModel.find({
+      subCategory: {
+        $in: [_id],
+      },
+    }).countDocuments();
+
+    if (checkProduct > 0) {
+      return res.status(400).json({
+        message: "Sub Category is already in use. Can't deleted",
+        error: true,
+        success: false,
+      });
+    }
 
     const deleteSub = await SubCategoryModel.findByIdAndDelete(_id)
